requirements.txt
numpy
pandas
matplotlib
flask
tensorflow
pyyaml
h5py
Pillow
opencv-python